replaced files, ATTACK_VERSION constant defined in bash
