+++ src/util/update_attack.sh
@@ -1,43 +1,32 @@
 ##
 ## This script creates a D3FEND ontology update from ATT&CK STIX JSON document
-## for any missing ATT&CK techniques and puts them under the correct parent
-## classes.
-##
-## After running this script, the developer needs to manually add:
-##  - Attack Tactics and relate them to the "Tactic-Technique" classes.
-##  - Remove any deprecated techniques as necessary.
+## After running the user must manually compare & replace d3fend-protege.updates.ttl
 ##
 
-pipenv run python src/util/update_attack.py || exit 1
+ATTACK_VERSION="13.1" # Current Attack Version 
 
-./bin/robot template \
-    --template reports/attack_update-missing_attack_ids-robot_template.csv \
-    --ontology-iri http://d3fend.mitre.org/ontologies/d3fend.owl \
-    --prefix "d3f: http://d3fend.mitre.org/ontologies/d3fend.owl#" \
-    -o src/ontology/_attack_update.ttl
+GREEN='\033[0;32m'
+YELLOW='\033[0;33m'
 
-# remove robot's uncessary type annotation
-# works on both bsd and gnu sed
-sed -e "s/\^\^xsd\:string//g" src/ontology/_attack_update.ttl > src/ontology/attack_update.ttl
+enterprise_attack="data/enterprise-attack-${ATTACK_VERSION}.json"
+if [ ! -f "$enterprise_attack" ]; then
+    echo -e "${GREEN}No enterprise attack file found"
+    echo -e "${GREEN}Running make download-attack \n"
+    make download-attack ATTACK_VERSION="${ATTACK_VERSION}"
+else 
+    echo -e "${GREEN}Using ${enterprise_attack} for attack data \n"
+fi
 
-./bin/robot merge \
-    --add-prefix "dcterms: http://purl.org/dc/terms/" \
-    --add-prefix "skos: http://www.w3.org/2004/02/skos/core#" \
-    -i src/ontology/d3fend-protege.ttl \
-    -i src/ontology/attack_update.ttl \
-    -o src/ontology/d3fend-protege.attack_update.ttl \
+cp src/ontology/d3fend-protege.ttl src/ontology/d3fend-protege.updates.ttl
 
-pipenv run ttlfmt src/ontology/d3fend-protege.attack_update.ttl
+pipenv run python src/util/test_cases.py  || exit 1
 
-echo Created new ontology file with updates here: src/ontology/d3fend-protege.attack_update.ttl
-echo Please manually review and compare to: src/ontology/d3fend-protege.ttl
+echo -e "${GREEN}All test cases passed \n"
 
+pipenv run python src/util/update_attack.py "$ATTACK_VERSION" || exit 1
 
-function cleanup(){
-    echo -e "\t deleting: $1"
-    rm -f $1
-}
+pipenv run ttlfmt src/ontology/d3fend-protege.updates.ttl
 
-echo cleaning up:
-cleanup src/ontology/_attack_update.ttl
-cleanup src/ontology/attack_update.ttl
+echo -e "${YELLOW}Created new ontology file with updates here: src/ontology/d3fend-protege.updates.ttl \n"
+echo -e "Please manually review and compare to: src/ontology/d3fend-protege.ttl \n"
+echo -e "If changes acceptable, replace files \n"
